Make styles !important.
I would be glad if you would pull this, otherwise I have to maintain my own copy.

I am aware that `!important` styles are bad practice but the thing is I am dealing with 3rd party code so my CSS needs to have !important rules in it to make sure elements are styled the way I define them.

With the native native browsers `object-fit` there is no problem because as object fit does not work like this polyfill, meaning it does not calculate and does this with margin and stuff so this values are ignored even if set with !important. One rule where this polyfill fails is `margin: 0 !important;` for example. I rather tried to make everything stick because I think when using this polyfill one really wants it to apply its styles no matter what. I think its not a bad Idea to do this even for other users so !important styles do never get into the way.

`setProperty` is IE9+ so it would fit.

This is currently untested and I only done this to the basic version only to see what you say.